refactor: 💡 RfhCarousel & ImageContainer
clean up the code
