+++ src/components/RfhCarousel/ImageContainer.tsx
@@ -1,0 +1,26 @@
+import { FC } from 'react'
+import { Box, BoxProps, Fade } from '@mui/material'
+
+export type ImageContainerProps = {
+  src?: HTMLImageElement['src']
+  maxHeight: number
+  ratio: number
+} & BoxProps
+
+export const ImageContainer: FC<ImageContainerProps> = ({ src, maxHeight, ratio, ...rest }) => (
+  <Box maxHeight={maxHeight} bgcolor={'black'} {...rest}>
+    <Fade in appear timeout={500}>
+      <img
+        src={src}
+        width={'100%'}
+        height={'100%'}
+        style={{
+          maxHeight: maxHeight,
+          objectFit: 'contain',
+          aspectRatio: ratio.toString(),
+        }}
+        loading={'lazy'}
+      />
+    </Fade>
+  </Box>
+)
