+++ src/components/RfhCarousel/RfhCarousel.tsx
@@ -1,9 +1,10 @@
 import { FC, useState } from 'react'
 import Stack, { StackProps } from '@mui/material/Stack'
 import { useMediaQuery, useTheme } from '@mui/material'
-import { ImageFrame } from './ImageFrame'
+import { ImageContainer } from './ImageContainer'
 import { ThumbsContainer } from './ThumbsContainer'
-import { CarouselImage, CarouselDefaults } from './config'
+import { CarouselDefaults } from './config'
+import type { CarouselImage } from './config'
 
 export type RfhCarouselProps = {
   images: CarouselImage[]
@@ -15,22 +16,20 @@ const RfhCarousel: FC<RfhCarouselProps> = ({ images, maxHeight, ratio, ...rest }
   const [currentImage, setCurrentImage] = useState<CarouselImage | undefined>(images[0])
   const theme = useTheme()
   const isXS = useMediaQuery(theme.breakpoints.only('xs'))
-  const aspectRatio = ratio || CarouselDefaults.ratio
-  const carouselImageMaxHeight = maxHeight || CarouselDefaults.maxHeight
 
   return (
-    <Stack
-      spacing={1.2}
-      {...rest}
-      style={{ aspectRatio: aspectRatio.toString(), maxHeight: carouselImageMaxHeight }}
-    >
-      <ImageFrame key={currentImage?.imageSrc} src={currentImage?.imageSrc} ratio={aspectRatio} />
+    <Stack spacing={CarouselDefaults.spacing} {...rest}>
+      <ImageContainer
+        src={currentImage?.imageSrc}
+        maxHeight={maxHeight || CarouselDefaults.maxHeight}
+        ratio={ratio || CarouselDefaults.ratio}
+      />
       <ThumbsContainer
+        images={images}
         options={{
-          thumbsGap: 1.2,
+          thumbsGap: CarouselDefaults.spacing,
           maxThumbsCount: isXS ? 5 : 6,
         }}
-        images={images}
         onThumbClick={(event, clickedImageIndex) => {
           setCurrentImage(prevState => {
             return images.find((item, index) => index === clickedImageIndex)
